+++ lib/zaikio/mission_control.rb
@@ -93,6 +93,8 @@ require "zaikio/mission_control/order_line_item"
 require "zaikio/mission_control/shipping_option"
 require "zaikio/mission_control/slot"
 require "zaikio/mission_control/production_frame"
+require "zaikio/mission_control/workstep"
+require "zaikio/mission_control/intermediate_product"
 
 module Zaikio
   module MissionControl